moved class into functions js
